refactor and redistribute between cold and client files
